+++ databases_preview.go
@@ -5,7 +5,7 @@ import (
 	stderrors "errors"
 
 	"github.com/Scalingo/go-scalingo/v9/debug"
-	"github.com/Scalingo/go-utils/errors/v2"
+	"github.com/Scalingo/go-utils/errors/v3"
 )
 
 var ErrDatabaseNotFound = stderrors.New("database not found")