+++ logs.go
@@ -5,9 +5,8 @@ import (
 	"net/http"
 	"net/url"
 
-	"gopkg.in/errgo.v1"
-
 	httpclient "github.com/Scalingo/go-scalingo/v9/http"
+	"github.com/Scalingo/go-utils/errors/v3"
 )
 
 type LogsService interface {
@@ -27,7 +26,7 @@ func (c *Client) LogsURL(ctx context.Context, app string) (*http.Response, error
 func (c *Client) Logs(ctx context.Context, logsURL string, n int, filter string) (*http.Response, error) {
 	u, err := url.Parse(logsURL)
 	if err != nil {
-		return nil, errgo.Mask(err)
+		return nil, errors.Wrap(ctx, err, "parse logs URL")
 	}
 	req := &httpclient.APIRequest{
 		NoAuth:   true,