+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * chore(databases): refactor constants
 * build(deps): update `github.com/golang-jwt/jwt` from v4 to v5
+* refactor: replace `errgo` with `github.com/Scalingo/go-utils/errors/v3`
 
 ## 9.2.0